Refactoring out customize activities into it's own menu option
