+++ app/build.gradle
@@ -46,6 +46,7 @@ android {
 }
 
 dependencies {
+    implementation 'androidx.activity:activity:1.8.0'
     androidTestImplementation "androidx.test:runner:1.5.2"
     androidTestImplementation "androidx.test:rules:1.5.0"
