+++ app/src/main/AndroidManifest.xml
@@ -16,6 +16,14 @@
         android:theme="@style/Theme.DefenseDrill"
         tools:targetApi="31">
         <activity
+            android:name=".ui.activities.CustomizeDatabaseActivity"
+            android:exported="false"
+            android:parentActivityName=".ui.activities.HomeActivity">
+            <meta-data
+                android:name="android.app.lib_name"
+                android:value="" />
+        </activity>
+        <activity
             android:name=".ui.activities.ViewAbstractCategoriesActivity"
             android:exported="false"
             android:parentActivityName=".ui.activities.HomeActivity">
