+++ app/src/main/res/layout/activity_customize_database.xml
@@ -6,54 +6,43 @@
 
     <LinearLayout xmlns:app="http://schemas.android.com/apk/res-auto"
         xmlns:tools="http://schemas.android.com/tools"
-        android:id="@+id/activityHome"
+        android:id="@+id/activityCustomizeDatabase"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:orientation="vertical"
-        tools:context=".ui.activities.HomeActivity" >
+        tools:context=".ui.activities.CustomizeDatabaseActivity" >
 
         <com.damienwesterman.defensedrill.ui.utils.TitleDescCard
-            android:id="@+id/generateDrillCard"
+            android:id="@+id/viewDrillsCard"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:layout_marginTop="8dp"
             android:layout_marginBottom="8dp"
             app:cardCornerRadius="8dp"
-            app:title="@string/generate_drill_title"
-            app:description="@string/generate_drills_description"
-            android:onClick="onCardClick"/>
-
-        <com.damienwesterman.defensedrill.ui.utils.TitleDescCard
-            android:id="@+id/customizeDatabaseCard"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="8dp"
-            android:layout_marginBottom="8dp"
-            app:cardCornerRadius="8dp"
-            app:title="@string/customize_database_title"
-            app:description="@string/customize_database_description"
-            android:onClick="onCardClick"/>
-
-        <com.damienwesterman.defensedrill.ui.utils.TitleDescCard
-            android:id="@+id/networkSetupCard"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="8dp"
-            android:layout_marginBottom="8dp"
-            app:cardCornerRadius="8dp"
-            app:title="@string/network_setup_title"
-            app:description="@string/network_setup_description"
+            app:title="@string/customize_drills_title"
+            app:description="@string/customize_drills_description"
             android:onClick="onCardClick" />
 
         <com.damienwesterman.defensedrill.ui.utils.TitleDescCard
-            android:id="@+id/feedbackCard"
+            android:id="@+id/viewCategoriesCard"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:layout_marginTop="8dp"
             android:layout_marginBottom="8dp"
             app:cardCornerRadius="8dp"
-            app:title="@string/feedback_title"
-            app:description="@string/feedback_description"
+            app:title="@string/customize_categories_title"
+            app:description="@string/customize_categories_description"
+            android:onClick="onCardClick" />
+
+        <com.damienwesterman.defensedrill.ui.utils.TitleDescCard
+            android:id="@+id/viewSubCategoriesCard"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:layout_marginTop="8dp"
+            android:layout_marginBottom="8dp"
+            app:cardCornerRadius="8dp"
+            app:title="@string/customize_sub_categories_title"
+            app:description="@string/customize_sub_categories_description"
             android:onClick="onCardClick" />
 
     </LinearLayout>
